Update and use newest version of views
