--- customer_relationship_management copy/customer_relationship/customer_relationship.view.yaml
@@ -1,12 +1,0 @@
-# Used by Airplane to identify your view. Do not change.
-slug: demo_customer_relationship
-
-# A human-readable name for your view.
-name: Customer relationship management dashboard
-
-# A human-readable description for your view.
-# description: "My Airplane view"
-
-# The path to the .view.tsx file containing the logic for this view. This
-# can be absolute or relative to the location of the definition file.
-entrypoint: customer_relationship.view.tsx
