+++ customer_relationship_management/package.json
@@ -3,13 +3,11 @@
   "version": "1.0.0",
   "license": "MIT",
   "dependencies": {
-    "@airplane/views": "^1.8.3",
-    "md5": "^2.3.0",
+    "@airplane/views": "v2.0.0-rc.5",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
   },
   "devDependencies": {
-    "@types/md5": "^2.3.2",
     "@types/react": "^18.0.17",
     "@types/react-dom": "^18.0.6",
     "typescript": "^4.8.2"
