+++ customer_relationship_management/tasks/create_lead.task.yaml
@@ -31,10 +31,6 @@ parameters:
     name: Address
     type: longtext
     required: true
-  - slug: opportunity_stage
-    name: Opportunity name
-    type: shorttext
-    required: true
 sql:
   # Replace with your own DB name
   resource: "[Demo DB]"
@@ -47,4 +43,4 @@ sql:
     country: "{{params.country}}"
     phone: "{{params.phone}}"
     address: "{{params.address}}"
-    opportunity_stage: "{{params.opportunity_stage}}"
+    opportunity_stage: "lead"
