+++ customer_relationship_management/tasks/create_touch_point_for_customer.task.yaml
