+++ customer_relationship_management/tasks/delete_customer.task.yaml
