+++ customer_relationship_management/tasks/delete_touch_point_for_customer.sql
