+++ customer_relationship_management/tasks/list_customer_touch_points.sql
