+++ customer_relationship_management/tasks/list_customers_by_stage.sql
