+++ customer_relationship_management/tasks/update_customer_stage.sql
